HT-60 HT-61  - geofield + leaflet + automatic geotagging seems to work quite nicely now! I've added 2 new fields to buildings:  1. geofield with leaflet map formatter + automatic geocoding  2. new geolocation address field, which is used for automated geo-location of buildings based on address.
The geolocation address field (text) is not displayed anywhere except on the building edit form. It is
used to store the geolocated address of the geofield. I tested prefilling this text field from the
street address (of space nodes from optime), and it seems to work surprisingly well even without defining a city etc.

So now updating the building data manually should be slightly more convenient, because of semi-automatic geolocationing.